Add hook for destructuring page response
